chore: port Ignition Plugin errors to Hardhat 3
Ignition errors in core are explicitly Ignition Errors. If one is
thrown, previously we would determine if it should be wrapped as a
Plugin Error and passed up.

We retain this mechanism but use HardhatError directly as Ignition is
core to Hardhat.

The goal in this commit was to port the `hardhat-ignition` error
messages directly and keep the wrapping behaviour as before.